+++ admin/src/main/scala/net/wiringbits/webapp/utils/ui/webTest/AdminView.scala
@@ -3,15 +3,15 @@ package net.wiringbits.webapp.utils.ui.webTest
 import io.github.nafg.simplefacade.Factory
 import japgolly.scalajs.react.vdom.VdomNode
 import net.wiringbits.webapp.utils.api.models.AdminGetTables
-import net.wiringbits.webapp.utils.ui.webTest.components.{EditGuesser, ListGuesser}
+import net.wiringbits.webapp.utils.ui.webTest.components._
 import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin
-import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.{Admin, simpleRestProvider}
+import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.{Admin, Resource, simpleRestProvider}
 import net.wiringbits.webapp.utils.ui.webTest.models.DataExplorerSettings
 import org.scalajs.macrotaskexecutor.MacrotaskExecutor.Implicits.global
 
 import scala.concurrent.Future
 
-private object AdminView {
+object AdminView {
   private def AdminTables(
       api: API,
       response: AdminGetTables.Response,
@@ -21,11 +21,10 @@ private object AdminView {
 
     def buildResources: List[VdomNode] = {
       response.data.map { table =>
-        reactadmin.Resource(
-          _.key := s"table-${table.name}",
+        Resource(
           _.name := table.name,
           _.list := ListGuesser(table),
-          _.edit := components.EditGuesser(table, dataExplorerSettings)
+          _.edit := EditGuesser(table, dataExplorerSettings)
         )
       }
     }
