+++ admin/src/main/scala/net/wiringbits/webapp/utils/ui/webTest/components/EditGuesser.scala
@@ -5,11 +5,10 @@ import japgolly.scalajs.react.React.Fragment
 import japgolly.scalajs.react.ScalaFnComponent
 import japgolly.scalajs.react.vdom.html_<^._
 import net.wiringbits.webapp.utils.api.models.AdminGetTables
-import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.ReactAdmin.useEditContext
-import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin
-import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.Edit
+import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin._
 import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.inputs._
-import net.wiringbits.webapp.utils.ui.webTest.models.{ButtonAction, ColumnType, DataExplorerSettings}
+import net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin.ReactAdmin.useEditContext
+import net.wiringbits.webapp.utils.ui.webTest.models._
 import net.wiringbits.webapp.utils.ui.webTest.utils.ResponseGuesser
 import org.scalajs.dom
 import org.scalajs.macrotaskexecutor.MacrotaskExecutor.Implicits.global
@@ -27,9 +26,12 @@ object EditGuesser {
     val inputs: List[VdomNode] = fields
       .map { fieldType =>
         fieldType.`type` match {
-          case ColumnType.Date => DateTimeInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
-          case ColumnType.Text => TextInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
-          case ColumnType.Email => TextInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
+          case ColumnType.Date =>
+            DateTimeInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
+          case ColumnType.Text =>
+            TextInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
+          case ColumnType.Email =>
+            TextInput(_.source := fieldType.name, _.disabled := fieldType.disabled)
           case ColumnType.Reference(reference, source) =>
             ReferenceInput(_.source := fieldType.name, _.reference := reference)(
               SelectInput(_.optionText := source, _.disabled := fieldType.disabled)
@@ -53,25 +55,25 @@ object EditGuesser {
     def buttons(ctx: js.Dictionary[js.Any]): List[VdomNode] = tableAction
       .map { x =>
         x.actions.map { action =>
-          reactadmin.Button(_.onClick := (() => onClick(action, ctx)))(action.text)
+          Button(_.onClick := (() => onClick(action, ctx)))(action.text)
         }: List[VdomNode]
       }
       .getOrElse(List.empty[VdomNode])
 
     val actions = ScalaFnComponent.withHooks[Unit].unchecked(useEditContext()).render { (_, ctx) =>
       {
-        reactadmin.TopToolbar()(buttons(ctx): _*)
+        TopToolbar()(buttons(ctx): _*)
       }: VdomNode
     }
 
-    val deleteButton: VdomNode = if (response.canBeDeleted) reactadmin.DeleteButton() else Fragment()
-    val toolbar: VdomNode = reactadmin.Toolbar()(
-      reactadmin.SaveButton(),
+    val deleteButton: VdomNode = if (response.canBeDeleted) DeleteButton() else Fragment()
+    val toolbar: VdomNode = Toolbar()(
+      SaveButton(),
       deleteButton
     )
 
-    reactadmin.Edit(_.actions := actions())(
-      reactadmin.SimpleForm(_.toolbar := toolbar)(inputs: _*)
+    Edit(_.actions := actions())(
+      SimpleForm(_.toolbar := toolbar)(inputs: _*)
     )
   }
 }
