+++ admin/src/main/scala/net/wiringbits/webapp/utils/ui/webTest/facades/reactadmin/EditButton.scala
@@ -1,0 +1,14 @@
+package net.wiringbits.webapp.utils.ui.webTest.facades.reactadmin
+
+import io.github.nafg.simplefacade.Implicits._
+import io.github.nafg.simplefacade._
+import japgolly.scalajs.react.vdom.VdomNode
+import net.wiringbits.webapp.utils.ui.webTest.facades
+
+import scala.scalajs.js
+
+object EditButton extends FacadeModule.Simple {
+  override def raw: js.Object = facades.reactadmin.ReactAdmin.EditButton
+  override def mkProps = new Props
+  class Props extends PropTypes
+}
